sonar fixes
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1735345 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/ss/util/ImageUtils.java
@@ -189,7 +189,7 @@ public class ImageUtils {
         if (isHSSF) {
             h *= 1 - anchor.getDy1()/256d;
         } else {
-            h -= anchor.getDy1()/EMU_PER_PIXEL;
+            h -= anchor.getDy1()/(double)EMU_PER_PIXEL;
         }
 
         while(h < scaledHeight){
